Refactor settings popover
